requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
djangorestframework==3.14.0
mysqlclient==2.2.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
